requirements.dev.txt
slackclient
slackeventsapi
flake8
